moved test to correct file
